pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "blaecktcpy"
version = "2.0.0"
description = "A Python TCP server for real-time streaming of named, typed signals in binary format"
readme = "README.md"
authors = [
    { name = "Sebastian Strobl" }
]
requires-python = ">=3.12"
license = "MIT"
keywords = ["blaeck", "tcp", "loggbok", "logging", "arduino"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: System :: Logging",
    "Topic :: Scientific/Engineering",
]

[project.urls]
Homepage = "https://github.com/sebaJoSt/blaecktcpy"
"Bug Reports" = "https://github.com/sebaJoSt/blaecktcpy/issues"
"Source" = "https://github.com/sebaJoSt/blaecktcpy"

[project.optional-dependencies]
serial = ["pyserial"]
